Add UI to Favorite/Unfavorite a Meal. Toast, too!
Can click the Favorite / Unfavorite button to change whether a Meal is in your Favorites list! A Toast notification renders to alert you of your choice. A patch request is also sent off to the backend, and your Favorite Meals list is updated! Only works for a hard-coded User ID so far. In the future we will check what the ID of the currently logged in User is and use that instead.